requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
httpx
pandas
pytest
passlib[bcrypt]
bcrypt<4
python-multipart
